backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
scikit-learn>=1.3,<1.7
pandas>=2.0,<2.3
numpy>=1.24,<2.0
mlflow>=2.10,<2.20
joblib>=1.3
pydantic>=2.5
imbalanced-learn>=0.11
httpx>=0.25
python-dotenv>=1.0
